app/requirements.txt
google-adk
yfinance
pydantic
praw
python-dotenv
litellm
streamlit>=1.31.0
requests>=2.31.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
flask 
flask-cors 
yfinance